+++ scripts/provision_kohya.sh
@@ -1,18 +1,22 @@
 #!/usr/bin/env bash
 set -eu
 
-rm -rf ${KOHYA_ROOT}
-
 # Install Kohya_ss
-git clone https://github.com/bmaltais/kohya_ss.git ${KOHYA_ROOT}
+[ -d ${KOHYA_ROOT} ] || git clone https://github.com/bmaltais/kohya_ss.git ${KOHYA_ROOT}
 
 cd ${KOHYA_ROOT}
 
-cp /config/kohya_ss/requirements* ./
-
+git fetch --tags
 git checkout ${KOHYA_VERSION}
 
-python3 -m venv --system-site-packages venv
+if [ -f install_complete ]; then 
+    echo "install_complete flag file found, skipping Kohya install..."
+    exit
+fi
+
+cp /config/kohya_ss/requirements* ./
+
+[ -d venv ] || python3 -m venv --system-site-packages venv
 source venv/bin/activate
 
 pip3 install --no-cache-dir torch==2.0.1 torchvision torchaudio --index-url https://download.pytorch.org/whl/cu118
@@ -25,5 +29,5 @@ pip3 install --no-cache-dir xformers==0.0.22 \
     tensorrt
 pip3 install -r requirements.txt
 pip3 install .
-pip3 cache purge
 deactivate
+touch install_complete
